+++ lib/parayaz/version.rb
@@ -1,3 +1,3 @@
 module Parayaz
-  VERSION = '1.0.4'
+  VERSION = '1.0.5'
 end
